+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.21.1
 - Fix: AddressAutocompleteField now supports `ui:placeholder` attribute
 - Fix: Upgrade to rjsf patch version for allOf support
+- Fix: Show error list on top, and scroll to top, on error
 
 - Chore: Speed up lambda by increasing memory from 128 MB -> 1024 MB